+++ core/services/ocr2/plugins/ccip/internal/ccipcommon/shortcuts.go
@@ -1,0 +1,54 @@
+package ccipcommon
+
+import (
+	"context"
+	"encoding/hex"
+	"fmt"
+
+	"github.com/ethereum/go-ethereum/common"
+	"golang.org/x/sync/errgroup"
+
+	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal"
+	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
+)
+
+func GetMessageIDsAsHexString(messages []internal.EVM2EVMMessage) []string {
+	messageIDs := make([]string, 0, len(messages))
+	for _, m := range messages {
+		messageIDs = append(messageIDs, "0x"+hex.EncodeToString(m.MessageId[:]))
+	}
+	return messageIDs
+}
+
+// GetDestinationTokens returns the destination chain fee tokens from the provided price registry
+// and the bridgeable tokens from the offramp.
+func GetDestinationTokens(ctx context.Context, offRamp ccipdata.OffRampReader, priceRegistry ccipdata.PriceRegistryReader) (fee, bridged []common.Address, err error) {
+	eg := new(errgroup.Group)
+
+	var destFeeTokens []common.Address
+	var destBridgeableTokens []common.Address
+
+	eg.Go(func() error {
+		tokens, err := priceRegistry.GetFeeTokens(ctx)
+		if err != nil {
+			return fmt.Errorf("get dest fee tokens: %w", err)
+		}
+		destFeeTokens = tokens
+		return nil
+	})
+
+	eg.Go(func() error {
+		tokens, err := offRamp.GetDestinationTokens(ctx)
+		if err != nil {
+			return fmt.Errorf("get dest bridgeable tokens: %w", err)
+		}
+		destBridgeableTokens = tokens
+		return nil
+	})
+
+	if err := eg.Wait(); err != nil {
+		return nil, nil, err
+	}
+
+	return destFeeTokens, destBridgeableTokens, nil
+}
