+++ core/services/ocr2/plugins/ccip/internal/ccipcommon/shortcuts_test.go
@@ -1,4 +1,4 @@
-package contractutil
+package ccipcommon
 
 import (
 	"math/rand"
