--- core/services/ocr2/plugins/ccip/internal/contractutil/shortcuts.go
@@ -1,15 +1,0 @@
-package contractutil
-
-import (
-	"encoding/hex"
-
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal"
-)
-
-func GetMessageIDsAsHexString(messages []internal.EVM2EVMMessage) []string {
-	messageIDs := make([]string, 0, len(messages))
-	for _, m := range messages {
-		messageIDs = append(messageIDs, "0x"+hex.EncodeToString(m.MessageId[:]))
-	}
-	return messageIDs
-}
